Add maven publishing support
Summary:
This adds preliminary support for publishing to maven by following https://our.intern.facebook.com/intern/dex/open-source-android/#maven-setup and tweaking some bits where it's outdated.

I've labelled this as 0.0.1, following semantic versioning (http://semver.org/).

Note that the javadoc publishing step fails after generating javadocs from the project itself because it can't find the dependencies (because javadoc is only aware of .jars, and not .aars, which are android dependencies). The work around is to apparently write code to extract the aar file and to then manually add them to the class path; I'll follow up with a second diff later for that ... for now the relevant javadocs will still be there in the generated jar file.

Reviewed By: xiphirx

Differential Revision: D6068032

fbshipit-source-id: 4baa92fe6c4eb4635c7e964805cef70b63779132