Add serializers for AppWakeupMetrics
Summary: Persisting to disk is very important specially for appwakeupmetrics since these usually restart the app in the background and the app then dies, so we lose most of the data if we don't save them to disk. Hence, implementing serializers for them as the first step. Will need to see if its easy to hook them into the app.

Reviewed By: kunalb

Differential Revision: D6179357

fbshipit-source-id: 371810b5c541dc76ad541ea0ed8ff928e4e6c7a0