Add VisibleToAvoidSynthetics annotation
Summary: Added a new annotation "VisibleToAvoidSynthetics", which is going to be used for those members whose default visibility has been set to package private instead of private, in order to prevent creation of synthetic accessors. This is to signal other contributors to not change their visibility to private. Also annotated some members with this annotation.