Fix failing tests in DeviceBatteryMetricsCollectorTest & AppWakeupMetricsCollectorTest
Summary: SystemMetricsLogger's delegate was being set to an anonymous class which was throwing a runtime exception, in the WakeLockMetricsCollectorTest class's setup method. Hence in the test suite, all the calls made to SystemMetricsLogger's wtf method were being rerouted to this delegate. This was leading to three tests failing in the DeviceBatteryMetricsCollectorTest & AppWakeupMetricsCollectorTest classes. In order to fix this, I deleted the lines of code which were setting the delegate in the setup method of WakeLockMetricsCollectorTest.