Add ability to set custom previous snapshot in StatefulSystemMetricsCollector
Summary:
This changes the API of StatefulSystemMetricsCollector slightly: the 3 argument explicit version no longer auto-initializes. I'll make sure to explicitly note this in the next version bump of the library to warn people realizing on this.

BluetoothMetrics was the only existing call site to use this, adjusted that in the code in this diff.

Differential Revision: D6583868

fbshipit-source-id: 377b62bef75fb94c8f8b31c957cfbc6284e7ac36