Treat invalid metrics as zeros for the sum of CompositeMetrics
Summary: When CompositeMetrics are deserialized, metrics that did not exist or that were invalid at the time of serialization will be marked as invalid. It is expected that, when a valid metric is then added to this invalid metric, that the invalid metrics will be treated as zeros and that the result will be valid.

Reviewed By: kunalb

Differential Revision: D10368456

fbshipit-source-id: bddf2da6256a0c78145594a53f86ea14955d7852