Add instrumentation for sensors to Battery Metrics.
Summary:
Tried to do a slightly simpler version of D15017990 that creates fewer classes and does less memory allocation: I only collect what we'll be publishing for now.

I haven't added support for attribution in the snapshots themselves yet, that'll be a follow up diff after hooking this up throughout.

Reviewed By: swkrueger

Differential Revision: D17190071

fbshipit-source-id: 132500d8a7163181acfef4068ea25fd91d42be3e