Make fbandroid/libraries/batterymetrics/serializers/src/main/java/com/facebook/battery/serializer/time:time nullsafe
Summary:
Mark TimeMetricsSerializer as nullsafe.

## General Context

This diff is part of an effort to make all of our Android apps nullsafe. This will not only help reduce NPEs but allow for smooth conversion from Java to Kotlin. If this is your team’s code, we would appreciate it if you could review our efforts as well as make the rest of your codebase nullsafe.

Reviewed By: jocelynluizzi13

Differential Revision: D59221943

fbshipit-source-id: 288352481512de2c12f6e8988299dfa5e89a2cbc